- Edits the style of the BookmarkSettings section - Adds an alert when the user tries to delete a bookmark

+++ src/components/settings/BookmarkSettings.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import getDomain from "../../services/getDomain";
+import { FiEdit, FiSave, FiTrash } from "react-icons/fi";
 
 const BookmarkSettings = ({ setPreventExit }) => {
   const [bookmarks, setBookmarks] = React.useState([]);
@@ -13,6 +14,10 @@ const BookmarkSettings = ({ setPreventExit }) => {
   const [showAdd, setShowAdd] = React.useState(false);
 
   function deleteBookmark(id) {
+    /* alert user to confirm deletion */
+    if (!window.confirm("Are you sure you want to delete this bookmark?")) {
+      return;
+    }
     const newBookmarks = bookmarks.filter((bookmark) => bookmark.id !== id);
     setBookmarks(newBookmarks);
     localStorage.setItem("bookmarks", JSON.stringify(newBookmarks));
@@ -53,6 +58,13 @@ const BookmarkSettings = ({ setPreventExit }) => {
     setBookmarks(newBookmarks);
     localStorage.setItem("bookmarks", JSON.stringify(newBookmarks));
     setEditId(null);
+    setNewBookmark({
+      id: "",
+      name: "",
+      url: "",
+      icon: null,
+    });
+
     setPreventExit(false);
   }
 
@@ -67,71 +79,73 @@ const BookmarkSettings = ({ setPreventExit }) => {
     }
   }, []);
   return (
-    <div className="mt-2 w-full">
-      <p className="text-sm text-gray-400">Bookmark Settings</p>
-      {bookmarks.map((bookmark) => (
-        <div key={bookmark.id} className="flex items-center mt-2 grow w-full">
-          <img
-            className="w-6 h-6 rounded-full"
-            src={`https://legal-white-orca.faviconkit.com/${getDomain(
-              bookmark.url
-            )}/32`}
-            alt={bookmark.name}
-          />
-          {editId !== bookmark.id && (
-            <div className="ml-2 w-full">
-              <h2 className="text-sm">{bookmark.name}</h2>
-              <p className="text-xs text-gray-400">{bookmark.url}</p>
-            </div>
-          )}
-          {editId === bookmark.id ? (
-            <div className="flex grow">
-              <p className="text-sm ml-1 mt-0.5">Name:</p>
-              <input
-                type="text"
-                name="name"
-                value={newBookmark.name}
-                onChange={handleBookmarkChange}
-                className="ml-1 text-sm w-80 border-b border-gray-500 bg-gray-700 rounded"
-              />
-              <p className="text-sm  ml-1 mt-0.5">URL:</p>
-              <input
-                type="text"
-                name="url"
-                value={newBookmark.url}
-                onChange={handleBookmarkChange}
-                className="ml-1 text-sm w-full border-b border-gray-500 bg-gray-700 rounded"
-              />
-              <button
-                className="ml-5 text-blue-500 ml-1"
-                onClick={() => editBookmark()}
-              >
-                Save
-              </button>
-            </div>
-          ) : (
-            <div className="flex justify-items-end">
-              <button
-                className="ml-auto text-blue-500 disabled:opacity-20"
-                onClick={() => {
-                  setEditId(bookmark.id);
-                  setNewBookmark(bookmark);
-                  setPreventExit(true);
-                }}
-                disabled={editId || showAdd}
-              >
-                Edit
-              </button>
-              <button
-                className="ml-auto text-red-500"
-                onClick={() => deleteBookmark(bookmark.id)}
-              >
-                Delete
-              </button>
-            </div>
-          )}
-        </div>
-      ))}
+    <div>
+      <div className="mt-2 w-full max-h-640 overflow-y-auto">
+        <p className="text-sm text-gray-400">Bookmark Settings</p>
+        {bookmarks.map((bookmark) => (
+          <div key={bookmark.id} className="flex items-center mt-2 grow w-full">
+            <img
+              className="w-6 h-6 rounded-full"
+              src={`https://legal-white-orca.faviconkit.com/${getDomain(
+                bookmark.url
+              )}/32`}
+              alt={bookmark.name}
+            />
+            {editId !== bookmark.id && (
+              <div className="ml-2 w-full">
+                <h2 className="text-sm">{bookmark.name}</h2>
+                <p className="text-xs text-gray-400">{bookmark.url}</p>
+              </div>
+            )}
+            {editId === bookmark.id ? (
+              <div className="flex grow">
+                <p className="text-sm ml-1 mt-2">Name:</p>
+                <input
+                  type="text"
+                  name="name"
+                  value={newBookmark.name}
+                  onChange={handleBookmarkChange}
+                  className="ml-1 text-sm w-80 border-b border-gray-500 bg-gray-700 rounded"
+                />
+                <p className="text-sm  ml-1 mt-2">URL:</p>
+                <input
+                  type="text"
+                  name="url"
+                  value={newBookmark.url}
+                  onChange={handleBookmarkChange}
+                  className="ml-1 text-sm w-full border-b border-gray-500 bg-gray-700 rounded"
+                />
+                <button
+                  className="ml-5 text-blue-500 ml-1"
+                  onClick={() => editBookmark()}
+                >
+                  <FiSave className="mx-auto" /> Save
+                </button>
+              </div>
+            ) : (
+              <div className="flex justify-items-end">
+                <button
+                  className="ml-auto text-blue-500 disabled:opacity-20"
+                  onClick={() => {
+                    setEditId(bookmark.id);
+                    setNewBookmark(bookmark);
+                    setPreventExit(true);
+                  }}
+                  disabled={editId || showAdd}
+                >
+                  <FiEdit className="w-5 h-5" />
+                </button>
+                <button
+                  className="ml-2 text-red-500"
+                  onClick={() => deleteBookmark(bookmark.id)}
+                >
+                  <FiTrash className="w-5 h-5" />
+                </button>
+              </div>
+            )}
+          </div>
+        ))}
+      </div>
       {!showAdd && (
         <div className="flex items-center mt-4 grow w-full">
           <div className="w-6 h-6 rounded-full bg-blue-500 text-white text-center">
